+++ forui/lib/widgets/alert.dart
@@ -5,4 +5,6 @@
 /// See https://forui.dev/docs/alert for working examples.
 library forui.widgets.alert;
 
-export '../src/widgets/alert/alert.dart' hide Variant;
+export '../src/widgets/alert/alert.dart' hide InheritedData, Variant;
+export '../src/widgets/alert/alert_icon.dart';
+export '../src/widgets/alert/alert_styles.dart';